feat(frontend): @agenta/settings + @agenta/settings-ui — the settings spine leaves the app
The tab model, the static-table helper and the page shell were already pure: no
antd, no @/oss. They move verbatim (git mv, so 111's featureFlags->preferences
rename keeps its history) and OSS keeps its three import paths as re-exports, so
none of the 13 tab files move.

Access is injected rather than read: useSettingsAccess needs edition detection,
project permissions and org state, none of which exist in a package. The package
owns the SettingsAccess shape and the visibility rules; each host computes the
flags. A host with no provider gets CLOSED_SETTINGS_ACCESS, so a missing provider
hides gated tabs instead of exposing them.
