+++ web/mobile/package.json
@@ -34,6 +34,8 @@
         "@agenta/sdk": "workspace:../packages/agenta-sdk",
         "@agenta/sessions": "workspace:../packages/agenta-sessions",
         "@agenta/sessions-ui": "workspace:../packages/agenta-sessions-ui",
+        "@agenta/settings": "workspace:../packages/agenta-settings",
+        "@agenta/settings-ui": "workspace:../packages/agenta-settings-ui",
         "@agenta/shared": "workspace:../packages/agenta-shared",
         "@agenta/ui": "workspace:../packages/agenta-ui",
         "@ai-sdk/react": "3.0.0-beta.153",
