+++ web/packages/agenta-settings-ui/package.json
@@ -1,0 +1,37 @@
+{
+    "name": "@agenta/settings-ui",
+    "version": "0.1.0",
+    "private": true,
+    "sideEffects": false,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "scripts": {
+        "build": "tsc --noEmit",
+        "types:check": "tsc --noEmit",
+        "lint": "eslint --config ./eslint.config.mjs src/",
+        "check": "pnpm run types:check && pnpm run lint"
+    },
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "dependencies": {
+        "@agenta/settings": "workspace:../agenta-settings",
+        "@agenta/shared": "workspace:../agenta-shared",
+        "@agenta/ui": "workspace:../agenta-ui",
+        "@phosphor-icons/react": "^2.1.10",
+        "clsx": "^2.1.1"
+    },
+    "peerDependencies": {
+        "jotai": ">=2.0.0",
+        "react": ">=18.0.0",
+        "next": ">=14.0.0"
+    },
+    "devDependencies": {
+        "@types/node": "^20.19.20",
+        "@types/react": "^19.0.10",
+        "jotai": "^2.15.0",
+        "react": "^19.0.0",
+        "typescript": "^5.9.3",
+        "next": "15.5.18"
+    }
+}
