+++ web/packages/agenta-settings-ui/src/index.ts
@@ -1,0 +1,4 @@
+export {
+    default as SettingsPageShell,
+    type SettingsPageShellProps,
+} from "./SettingsPageShell"
