+++ web/packages/agenta-settings/package.json
@@ -1,0 +1,33 @@
+{
+    "name": "@agenta/settings",
+    "version": "0.1.0",
+    "private": true,
+    "sideEffects": false,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "scripts": {
+        "build": "tsc --noEmit",
+        "types:check": "tsc --noEmit",
+        "lint": "eslint --config ./eslint.config.mjs src/",
+        "test": "pnpm run test:unit",
+        "test:unit": "vitest run",
+        "check": "pnpm run types:check && pnpm run lint"
+    },
+    "exports": {
+        ".": "./src/index.ts"
+    },
+    "dependencies": {
+        "@agenta/shared": "workspace:../agenta-shared",
+        "@agenta/ui": "workspace:../agenta-ui"
+    },
+    "peerDependencies": {
+        "react": ">=18.0.0"
+    },
+    "devDependencies": {
+        "@types/node": "^20.19.20",
+        "@types/react": "^19.0.10",
+        "react": "^19.0.0",
+        "typescript": "^5.9.3",
+        "vitest": "^4.1.4"
+    }
+}
