+++ web/packages/agenta-settings/src/index.ts
@@ -1,0 +1,7 @@
+export * from "./navigation"
+export {useStaticTable} from "./useStaticTable"
+export {
+    SettingsAccessProvider,
+    useSettingsAccess,
+    CLOSED_SETTINGS_ACCESS,
+} from "./access"
